refactor: breadcrumbs to shared
