+++ src/app/components/index.ts
@@ -1,5 +1,4 @@
 export { Actions } from './actions';
-export { Breadcrumbs } from './breadcrumbs';
 export { CourseCard } from './course-card';
 export { DateLang } from './date-lang';
 export { EventCard, type EventCardProps } from './event-card';
