+++ src/pages/javascript-ru.tsx
@@ -1,4 +1,3 @@
-import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { AboutVideo } from '@/features/about-video';
 import { Certification } from '@/features/certification';
@@ -7,6 +6,7 @@ import { CourseMain } from '@/features/course-main';
 import { Required } from '@/features/required';
 import { StudyPath } from '@/features/study-path';
 import { TrainingProgram } from '@/features/training-program';
+import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 const COURSE_NAME = 'js / front-end ru';
