+++ src/pages/nodejs.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react';
-import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
@@ -8,6 +7,7 @@ import { Required } from '@/features/required';
 import { Trainers } from '@/features/trainers';
 import { nodejs } from '@/features/trainers/nodejs.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 const COURSE_NAME = 'node.js';
