+++ src/pages/react.tsx
@@ -1,4 +1,3 @@
-import { Breadcrumbs } from '@/app/components';
 import { About } from '@/features/about';
 import { Certification } from '@/features/certification';
 import { Communication } from '@/features/communication';
@@ -6,6 +5,7 @@ import { CourseMain } from '@/features/course-main';
 import { Trainers } from '@/features/trainers';
 import { reactEn } from '@/features/trainers/react-en.data.ts';
 import { TrainingProgram } from '@/features/training-program';
+import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 const COURSE_NAME = 'react';
