+++ src/pages/school.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react';
-import { Breadcrumbs } from '@/app/components';
 import { useTitle } from '@/app/hooks';
 import { Principles } from '@/features/principles';
 import {
@@ -12,6 +11,7 @@ import {
   Requirements,
   StudyWithUs,
 } from '@/features/school';
+import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 export const School: FC = () => {
   useTitle('The Rolling Scopes School');
