+++ src/shared/ui/breadcrumbs/breadcrumbs.scss
@@ -8,7 +8,7 @@
       padding: 24px 15px 10px;
     }
 
-    padding: 24px 120px 10px 120px;
+    padding: 24px 120px 10px;
 
     ul {
       display: flex;
